extension completed

+++ src/main/java/com/booleanuk/api/cinema/controllers/CustomerController.java
@@ -2,6 +2,10 @@ package com.booleanuk.api.cinema.controllers;
 
 import com.booleanuk.api.cinema.models.Customer;
 import com.booleanuk.api.cinema.repositories.CustomerRepository;
+import com.booleanuk.api.cinema.response.CustomerListResponse;
+import com.booleanuk.api.cinema.response.CustomerResponse;
+import com.booleanuk.api.cinema.response.ErrorResponse;
+import com.booleanuk.api.cinema.response.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,32 +21,61 @@ public class CustomerController {
     private CustomerRepository customerRepository;
 
     @GetMapping
-    public List<Customer> getAllCustomers() {
-        return this.customerRepository.findAll();
+    public ResponseEntity<CustomerListResponse> getAllCustomers() {
+        CustomerListResponse customerListResponse = new CustomerListResponse();
+        customerListResponse.set(this.customerRepository.findAll());
+        return ResponseEntity.ok(customerListResponse);
     }
 
     @PostMapping
-    public ResponseEntity<Customer> addCustomer(@RequestBody Customer customer) {
-        customer.setCreatedAt(java.time.LocalDateTime.now());
-        customer.setUpdatedAt(null);
-        return new ResponseEntity<>(this.customerRepository.save(customer), HttpStatus.CREATED);
+    public ResponseEntity<Response<?>> addCustomer(@RequestBody Customer customer) {
+        CustomerResponse customerResponse = new CustomerResponse();
+        try {
+            customerResponse.set(this.customerRepository.save(customer));
+        } catch (Exception e) {
+            ErrorResponse errorResponse = new ErrorResponse();
+            errorResponse.set("bad request");
+            return new ResponseEntity<>(errorResponse,HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity<>(customerResponse,HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Customer> updateCustomer(@RequestBody Customer customer, @PathVariable int id) {
-        Customer updatedCustomer = this.customerRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Customer Not Found"));
+    public ResponseEntity<Response<?>> updateCustomer(@RequestBody Customer customer, @PathVariable int id) {
+        Customer updatedCustomer = null;
+        try {
+            updatedCustomer = this.customerRepository.findById(id).orElse(null);
+        } catch (Exception e) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Bad Request");
+            return new ResponseEntity<>(error,HttpStatus.BAD_REQUEST);
+        }
+        if (updatedCustomer == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Not Found");
+            return new ResponseEntity<>(error,HttpStatus.NOT_FOUND);
+        }
         updatedCustomer.setName(customer.getName());
         updatedCustomer.setEmail(customer.getEmail());
         updatedCustomer.setPhone(customer.getPhone());
         updatedCustomer.setUpdatedAt(java.time.LocalDateTime.now());
-        this.customerRepository.save(updatedCustomer);
-        return ResponseEntity.ok(updatedCustomer);
+        updatedCustomer = this.customerRepository.save(updatedCustomer);
+        CustomerResponse customerResponse = new CustomerResponse();
+        customerResponse.set(updatedCustomer);
+        return new ResponseEntity<>(customerResponse,HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Customer> deleteCustomer(@PathVariable int id) {
-        Customer customer = this.customerRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Customer Not Found"));
+    public ResponseEntity<Response<?>> deleteCustomer(@PathVariable int id) {
+        Customer customer = this.customerRepository.findById(id).orElse(null);
+        if (customer == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Customer Not Found");
+            return new ResponseEntity<>(error,HttpStatus.NOT_FOUND);
+        }
         this.customerRepository.delete(customer);
-        return ResponseEntity.ok(customer);
+        CustomerResponse customerResponse = new CustomerResponse();
+        customerResponse.set(customer);
+        return ResponseEntity.ok(customerResponse);
     }
 }
